requirements.txt
flask
flask_sqlalchemy
flask_login
itsdangerous
python-dotenv
openai
requests
gunicorn